+++ r-private-endpoint.tf
@@ -4,14 +4,27 @@ resource "azurerm_private_endpoint" "private_endpoint" {
 
   resource_group_name = var.resource_group_name
 
+  custom_network_interface_name = var.custom_private_endpoint_nic_name
+
   subnet_id = var.subnet_id
 
-  dynamic "private_dns_zone_group" {
-    for_each = local.is_not_private_link_service ? ["private_dns_zone_group"] : []
-
+  dynamic "ip_configuration" {
+    for_each = var.ip_configurations
     content {
-      name                 = local.private_dns_zone_group_name
-      private_dns_zone_ids = var.use_existing_private_dns_zones ? var.private_dns_zones_ids : [for zone in module.private_dns_zones : zone.private_dns_zone_id]
+      name               = ip_configuration.value.name
+      member_name        = local.is_not_private_link_service ? ip_configuration.value.member_name : null
+      subresource_name   = local.is_not_private_link_service ? coalesce(ip_configuration.value.subresource_name, var.subresource_name) : null
+      private_ip_address = ip_configuration.value.private_ip_address
+    }
+  }
+
+  dynamic "private_dns_zone_group" {
+    for_each = local.is_not_private_link_service ? ["enabled"] : []
+    content {
+      name = local.private_dns_zone_group_name
+      private_dns_zone_ids = var.use_existing_private_dns_zones ? var.private_dns_zones_ids : [
+        for zone in module.private_dns_zones : zone.private_dns_zone_id
+      ]
     }
   }