+++ variables.tf
@@ -33,6 +33,25 @@ variable "subnet_id" {
   type        = string
 }
 
+variable "ip_configurations" {
+  description = <<EOD
+List of IP Configuration object. Any modification to the parameters of the IP Configuration object forces a new resource to be created.
+```
+name               = Name of the IP Configuration.
+member_name        = Member name of the IP Configuration. If it is not specified, it will use the value of `subresource_name`. Only valid if `target_resource` is not a Private Link Service.
+subresource_name   = Subresource name of the IP Configuration. Only valid if `target_resource` is not a Private Link Service.
+private_ip_address = Private IP address within the Subnet of the Private Endpoint.
+```
+EOD
+  type = list(object({
+    name               = optional(string, "default")
+    member_name        = optional(string)
+    subresource_name   = optional(string)
+    private_ip_address = string
+  }))
+  default = []
+}
+
 variable "is_manual_connection" {
   description = "Does the Private Endpoint require manual approval from the remote resource owner? Default to `false`."
   type        = bool